// src/content/certificateFiles.generated.ts
export interface CertificateFile {
  title: string;
  category: string;
  categorySlug: string;
  fileName: string;
  type: "pdf" | "image";
  url: string;
}